Change the type of number variables from int to size_t. The type of order variables remains int. The casting from unsigned long to int only occurs for the calculation of order. It won't be a problem because the order cannot be that high.

+++ src/core/Input.h
@@ -47,7 +47,7 @@ struct ConfigFiber {
   Material material;
   double radius;
   int N_max;
-  int P;
+  size_t P;
 };
 struct Fiber {
   std::string configID;
